# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lumina HQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threads:
    get:
      operationId: getThreads
      tags: [threads]
      summary: Get all synced email threads
      parameters:
        - in: query
          name: classification
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Thread list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadsResponse"
  /threads/{id}:
    get:
      operationId: getThread
      tags: [threads]
      summary: Get single thread
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadResponse"
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List all active suppliers
      parameters:
        - in: query
          name: includeInactive
          schema:
            type: boolean
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Suppliers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppliersResponse"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier with categories
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
    delete:
      operationId: deactivateSupplier
      tags: [suppliers]
      summary: Deactivate a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /suppliers/{id}/categories:
    post:
      operationId: addSupplierCategory
      tags: [suppliers]
      summary: Add a product category to a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierCategoryResponse"
  /suppliers/{id}/categories/{catId}:
    delete:
      operationId: removeSupplierCategory
      tags: [suppliers]
      summary: Remove a category from a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: catId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /suppliers/{id}/performance:
    get:
      operationId: getSupplierPerformance
      tags: [suppliers]
      summary: Get supplier performance metrics
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPerformanceResponse"
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    EmailThread:
      type: object
      required: [id, threadId, subject, senderName, senderEmail, receivedAt, isRfq, syncedAt]
      properties:
        id:
          type: integer
        threadId:
          type: string
        subject:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
        receivedAt:
          type: string
        snippet:
          type: ["string", "null"]
        bodyText:
          type: ["string", "null"]
        classification:
          type: ["string", "null"]
        isRfq:
          type: boolean
        hasAttachments:
          type: boolean
        attachmentParsed:
          type: boolean
        attachmentType:
          type: ["string", "null"]
        syncedAt:
          type: string
    ThreadsResponse:
      type: object
      required: [threads]
      properties:
        threads:
          type: array
          items:
            $ref: "#/components/schemas/EmailThread"
    ThreadResponse:
      type: object
      required: [thread]
      properties:
        thread:
          $ref: "#/components/schemas/EmailThread"
...
    SupplierCategoryRecord:
      type: object
      required: [id, supplierId, category, isPreferred]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        category:
          type: string
        isPreferred:
          type: boolean
        notes:
          type: ["string", "null"]
    SupplierRecord:
      type: object
      required: [id, name, company, email, country, currency, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: string
        currency:
          type: string
        typicalLeadTimeDays:
          type: ["integer", "null"]
        typicalResponseTimeHours:
          type: ["integer", "null"]
        paymentTerms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    SupplierWithCategories:
      allOf:
        - $ref: "#/components/schemas/SupplierRecord"
        - type: object
          required: [categories]
          properties:
            categories:
              type: array
              items:
                $ref: "#/components/schemas/SupplierCategoryRecord"
    SuppliersResponse:
      type: object
      required: [suppliers]
      properties:
        suppliers:
          type: array
          items:
            $ref: "#/components/schemas/SupplierWithCategories"
    SupplierDetailResponse:
      type: object
      required: [supplier]
      properties:
        supplier:
          $ref: "#/components/schemas/SupplierWithCategories"
    SupplierInput:
      type: object
      required: [name, company, email]
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        currency:
          type: string
        typicalLeadTimeDays:
          type: integer
        typicalResponseTimeHours:
          type: integer
        paymentTerms:
          type: string
        notes:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/SupplierCategoryInput"
    SupplierUpdateInput:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        currency:
          type: string
        typicalLeadTimeDays:
          type: integer
        typicalResponseTimeHours:
          type: integer
        paymentTerms:
          type: string
        notes:
          type: string
        isActive:
          type: boolean
    SupplierCategoryInput:
      type: object
      required: [category]
      properties:
        category:
          type: string
        isPreferred:
          type: boolean
        notes:
          type: string
    SupplierCategoryResponse:
      type: object
      required: [category]
      properties:
        category:
          $ref: "#/components/schemas/SupplierCategoryRecord"
    SupplierPerformanceMetrics:
      type: object
      required: [totalRfqs, totalContacted, totalResponded, totalSelected, responseRatePercent, selectionRatePercent]
      properties:
        totalRfqs:
          type: integer
        totalContacted:
          type: integer
        totalResponded:
          type: integer
        totalSelected:
          type: integer
        responseRatePercent:
          type: integer
        selectionRatePercent:
          type: integer
        avgResponseTimeHours:
          type: ["number", "null"]
    SupplierPerformanceResponse:
      type: object
      required: [performance, history]
      properties:
        performance:
          $ref: "#/components/schemas/SupplierPerformanceMetrics"
        history:
          type: array
          items:
            type: object
...