# lib/api-zod/src/generated/types/emailThread.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumina HQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailThread {
  id: number;
  threadId: string;
  subject: string;
  senderName: string;
  senderEmail: string;
  receivedAt: string;
  /** @nullable */
  snippet?: string | null;
  /** @nullable */
  bodyText?: string | null;
  /** @nullable */
  classification?: string | null;
  isRfq: boolean;
  hasAttachments?: boolean;
  attachmentParsed?: boolean;
  /** @nullable */
  attachmentType?: string | null;
  syncedAt: string;
}