# lib/api-zod/src/generated/types/supplierPerformanceMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumina HQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierPerformanceMetrics {
  totalRfqs: number;
  totalContacted: number;
  totalResponded: number;
  totalSelected: number;
  responseRatePercent: number;
  selectionRatePercent: number;
  /** @nullable */
  avgResponseTimeHours?: number | null;
}