# lib/api-zod/src/generated/types/threadsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumina HQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailThread } from './emailThread';

export interface ThreadsResponse {
  threads: EmailThread[];
}